sms_success

+++ src/main/java/com/alcohol/sul/board/qna/QnaDTO.java
@@ -6,7 +6,7 @@ import com.alcohol.sul.board.BoardDTO;
 
 public class QnaDTO extends BoardDTO{
 	private QnaDTO qnaDTO;
-	private Integer sms;
+	private Long sms;
 	private Long ref;
 	private Long step;
 	private Long depth;
@@ -43,10 +43,10 @@ public class QnaDTO extends BoardDTO{
 	public void setDepth(Long depth) {
 		this.depth = depth;
 	}
-	public Integer getSms() {
+	public Long getSms() {
 		return sms;
 	}
-	public void setSms(Integer sms) {
+	public void setSms(Long sms) {
 		this.sms = sms;
 	}
 }
